policies for discount

+++ modules/Hadihosseini88/Discount/src/Http/Controllers/DiscountController.php
@@ -14,6 +14,7 @@ class DiscountController extends Controller
 {
     public function index(CourseRepo $courseRepo, DiscountRepo $discountRepo)
     {
+        $this->authorize('manage', Discount::class);
         $discounts = $discountRepo->paginateAll();
         $courses = $courseRepo->getAll(Course::CONFIRMATION_STATUS_ACCEPTED);
         return view('Discount::index', compact('courses', 'discounts'));
@@ -21,21 +22,23 @@ class DiscountController extends Controller
 
     public function store(DiscountRequest $request, DiscountRepo $repo)
     {
+        $this->authorize('manage', Discount::class);
         $repo->store($request->all());
-
         newFeedback();
-
         return back();
     }
 
     public function edit(Discount $discount, CourseRepo $courseRepo)
     {
+        $this->authorize('manage', Discount::class);
         $courses = $courseRepo->getAll(Course::CONFIRMATION_STATUS_ACCEPTED);
         return view('Discount::edit', compact('discount', 'courses'));
     }
 
     public function update(Discount $discount, DiscountRequest $request, DiscountRepo $repo)
     {
+
+        $this->authorize('manage', Discount::class);
         $repo->update($discount->id, $request->all());
         newFeedback();
         return redirect(route('discounts.index'));
@@ -44,8 +47,8 @@ class DiscountController extends Controller
 
     public function destroy(Discount $discount)
     {
+        $this->authorize('manage', Discount::class);
         $discount->delete();
-
         return AjaxResponses::SuccessResponse();
     }
